Исправляет функцию отрисовки объявления на странице
